Enable/disable flycheck checker more idiomatically
Flycheck can determine for itself whether the dante checker should be
enabled, by supplying it with a `:predicate` function that returns t
if and only dante-mode is active.

The dante checker is then simply added to the list of
globally-available checkers, so that it works the same way as every
other checker.

With this change, disabling dante-mode simply causes the next-best
Haskell checker to be selected until dante-mode is re-enabled, which
is the expected behaviour, while the user can still disable
flycheck-mode in the buffer to stop checking entirely.