Support flycheck's ability to show "other file" errors
Dante's flychecker currently reports "no errors" in the case where a
module cannot even be compiled because it depends on another broken
source file. This leads to mistrust of the checker because
clearly-invalid syntax can appear to be considered valid.

For a while now Flycheck has had a mechanism for allowing such
"upstream" errors to be nominally reported in the current buffer (on
the first line) if they are at least
`flycheck-relevant-error-other-file-minimum-level`. All that is
necessary for checkers to support this is for them to parse and return
errors for the current buffer where the `:filename` is a different
file.

This commit makes such a (minimal) change to the Dante checker,
bringing its behaviour into line with comparable checkers such as the
Elm and Intero checkers.

(Note that if they prefer, users can disable "other file" errors by
setting the above variable to nil.)