Convert forward slashes to backslashes when canonicalizing on Windows
On Windows, filenames are sometimes reported with backslashes and other
times reported with forward slashes. Canonicalization should ideally
convert all slashes to backslashes on Windows to avoid this problem.
This commit implements this change.

Note that on non-Windows systems, canonicalization should not similarly
convert all slashes to forward slashes. This is because non-Windows
systems may accept backslashes as part of file names, and these
backslashes should not be converted to forward slashes.