Emphasized dante-tap-type-time in README
I think it is a really cool feature, but it is really hard to find out about it, you have to end up reading the very last sentence in the README, and it is kind of mentioned as a side note. Therefore I tried to put a mention of it on more prominent place.